define C API
