+++ include/expanding.h
@@ -1,48 +1,45 @@
 #pragma once
 
-#include "export.h"
 #include "grouped_array.h"
+#include "rolling.h"
+#include "stats.h"
 
-extern "C" {
-DLL_EXPORT int
-GroupedArrayFloat32_ExpandingMeanTransform(GroupedArrayHandle handle, int lag,
-                                           float *out, float *agg);
-DLL_EXPORT int
-GroupedArrayFloat64_ExpandingMeanTransform(GroupedArrayHandle handle, int lag,
-                                           double *out, double *agg);
+template <typename T>
+inline void ExpandingMeanTransform(const T *data, int n, T *out, T *agg) {
+  T accum = static_cast<T>(0.0);
+  for (int i = 0; i < n; ++i) {
+    accum += data[i];
+    out[i] = accum / (i + 1);
+  }
+  *agg = static_cast<T>(n);
+}
 
-DLL_EXPORT int
-GroupedArrayFloat32_ExpandingStdTransform(GroupedArrayHandle handle, int lag,
-                                          float *out, float *agg);
-DLL_EXPORT int
-GroupedArrayFloat64_ExpandingStdTransform(GroupedArrayHandle handle, int lag,
-                                          double *out, double *agg);
+template <typename T>
+inline void ExpandingStdTransform(const T *data, int n, T *out, T *agg) {
+  RollingStdTransformWithStats(data, n, out, agg, true, n, 2);
+}
 
-DLL_EXPORT int
-GroupedArrayFloat32_ExpandingMinTransform(GroupedArrayHandle handle, int lag,
-                                          float *out);
-DLL_EXPORT int
-GroupedArrayFloat64_ExpandingMinTransform(GroupedArrayHandle handle, int lag,
-                                          double *out);
+template <typename T> struct ExpandingMinTransform {
+  void operator()(const T *data, int n, T *out) {
+    RollingCompTransform(std::less<T>(), data, n, out, n, 1);
+  }
+};
 
-DLL_EXPORT int
-GroupedArrayFloat32_ExpandingMaxTransform(GroupedArrayHandle handle, int lag,
-                                          float *out);
-DLL_EXPORT int
-GroupedArrayFloat64_ExpandingMaxTransform(GroupedArrayHandle handle, int lag,
-                                          double *out);
+template <typename T> struct ExpandingMaxTransform {
+  void operator()(const T *data, int n, T *out) {
+    RollingCompTransform(std::greater<T>(), data, n, out, n, 1);
+  }
+};
 
-DLL_EXPORT int
-GroupedArrayFloat32_ExpandingQuantileTransform(GroupedArrayHandle handle,
-                                               int lag, float p, float *out);
-DLL_EXPORT int
-GroupedArrayFloat64_ExpandingQuantileTransform(GroupedArrayHandle handle,
-                                               int lag, double p, double *out);
+template <typename T>
+inline void ExpandingQuantileTransform(const T *data, int n, T *out, T p) {
+  RollingQuantileTransform(data, n, out, n, 1, p);
+}
 
-DLL_EXPORT int
-GroupedArrayFloat32_ExpandingQuantileUpdate(GroupedArrayHandle handle, int lag,
-                                            float p, float *out);
-DLL_EXPORT int
-GroupedArrayFloat64_ExpandingQuantileUpdate(GroupedArrayHandle handle, int lag,
-                                            double p, double *out);
+template <typename T>
+inline void ExpandingQuantileUpdate(const T *data, int n, T *out, T p) {
+  T *buffer = new T[n];
+  std::copy(data, data + n, buffer);
+  *out = Quantile(buffer, p, n);
+  delete[] buffer;
 }
