+++ include/grouped_array.h
@@ -4,8 +4,6 @@
 #include <cmath>
 #include <cstdint>
 
-#include "export.h"
-
 using GroupedArrayHandle = void *;
 using indptr_t = int32_t;
