+++ include/grouped_array_functions.h
@@ -1,47 +1,29 @@
-#include "export.h"
 #include "grouped_array.h"
 
-extern "C" {
-DLL_EXPORT int GroupedArrayFloat32_Create(const float *data, indptr_t n_data,
-                                          indptr_t *indptr, indptr_t n_indptr,
-                                          int num_threads,
-                                          GroupedArrayHandle *out);
-DLL_EXPORT int GroupedArrayFloat64_Create(const double *data, indptr_t n_data,
-                                          indptr_t *indptr, indptr_t n_indptr,
-                                          int num_threads,
-                                          GroupedArrayHandle *out);
+template <typename T>
+inline void IndexFromEnd(const T *data, int n, T *out, int k) {
+  if (k > n) {
+    *out = std::numeric_limits<T>::quiet_NaN();
+  } else {
+    *out = data[n - 1 - k];
+  }
+}
 
-DLL_EXPORT int GroupedArrayFloat32_Delete(GroupedArrayHandle handle);
-DLL_EXPORT int GroupedArrayFloat64_Delete(GroupedArrayHandle handle);
+template <typename T> inline void Head(const T *data, int n, T *out, int k) {
+  int m = std::min(k, n);
+  std::copy(data, data + m, out);
+  std::fill(out + m, out + k, std::numeric_limits<T>::quiet_NaN());
+}
 
-DLL_EXPORT int GroupedArrayFloat32_IndexFromEnd(GroupedArrayHandle handle,
-                                                int k, float *out);
-DLL_EXPORT int GroupedArrayFloat64_IndexFromEnd(GroupedArrayHandle handle,
-                                                int k, double *out);
+template <typename T> inline void Tail(const T *data, int n, T *out, int k) {
+  int m = std::min(k, n);
+  std::fill(out, out + k - m, std::numeric_limits<T>::quiet_NaN());
+  std::copy(data + n - m, data + n, out + k - m);
+}
 
-DLL_EXPORT void GroupedArrayFloat32_Head(GroupedArrayHandle handle, int k,
-                                         float *out);
-DLL_EXPORT void GroupedArrayFloat64_Head(GroupedArrayHandle handle, int k,
-                                         double *out);
-
-DLL_EXPORT void GroupedArrayFloat32_Tail(GroupedArrayHandle handle, int k,
-                                         float *out);
-DLL_EXPORT void GroupedArrayFloat64_Tail(GroupedArrayHandle handle, int k,
-                                         double *out);
-
-DLL_EXPORT void GroupedArrayFloat32_Tails(GroupedArrayHandle handle,
-                                          const indptr_t *indptr_out,
-                                          float *out);
-DLL_EXPORT void GroupedArrayFloat64_Tails(GroupedArrayHandle handle,
-                                          const indptr_t *indptr_out,
-                                          double *out);
-
-DLL_EXPORT void GroupedArrayFloat32_Append(GroupedArrayHandle handle,
-                                           GroupedArrayHandle other_handle,
-                                           const indptr_t *out_indptr,
-                                           float *out_data);
-DLL_EXPORT void GroupedArrayFloat64_Append(GroupedArrayHandle handle,
-                                           GroupedArrayHandle other_handle,
-                                           const indptr_t *out_indptr,
-                                           double *out_data);
+template <typename T>
+inline void Append(const T *data, int n, const T *other_data, int other_n,
+                   T *out) {
+  std::copy(data, data + n, out);
+  std::copy(other_data, other_data + other_n, out + n);
 }
