+++ include/kpss.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "export.h"
 #include "stats.h"
 
 #include <cmath>
