+++ include/lag.h
@@ -1,11 +1,5 @@
 #pragma once
 
-#include "export.h"
-#include "grouped_array.h"
-
-extern "C" {
-DLL_EXPORT int GroupedArrayFloat32_LagTransform(GroupedArrayHandle handle,
-                                                int lag, float *out);
-DLL_EXPORT int GroupedArrayFloat64_LagTransform(GroupedArrayHandle handle,
-                                                int lag, double *out);
+template <typename T> inline void LagTransform(const T *data, int n, T *out) {
+  std::copy(data, data + n, out);
 }
